Class merge + contributors update
